fix(remarks) вынес работу с запросами
